bots/formatter/requirements.txt
zulip>=0.9.0
PyYAML>=6.0
